New whitelist system + stickyrole system

+++ helper_cogs/mod.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, List
+from itertools import chain
+from typing import Callable, Tuple, Set, TYPE_CHECKING
 
 from fuzzy_match import match
 
@@ -13,7 +14,7 @@ from utils import BotEmojis
 
 if TYPE_CHECKING:
     from helper_bot import NotGDKID
-
+    from utils import NGKContext
 
 muted = lambda user: f"""
 hello, i have been summoned by the great `{user.name}` to restore some peace and sanity in this hell of a server.
@@ -33,6 +34,12 @@ if you wanna get unmuted, dm `gdkid#0111` so he can laugh at you {BotEmojis.HAHA
 this message is also able to be seen by the entire server, because it really is just that funny
 """
 
+insert_q: Callable[[int], str] = lambda stop: \
+    """INSERT INTO stickyroles VALUES {0}
+    ON CONFLICT ON CONSTRAINT stickyroles_pkey
+    DO NOTHING""".format(",".join(
+        f"(${i}, ${i + 1})" for i in range(1, stop + 1, 2)
+    ))
 
 class Mod(commands.Cog):
     def __init__(self, client: NotGDKID) -> None:
@@ -48,13 +55,16 @@ class Mod(commands.Cog):
     async def cog_unload(self) -> None:
         self.client.tree.remove_command("shut the fuck up")
 
+    async def cog_check(self, ctx: NGKContext):
+        return ctx.guild and ctx.guild.id == self.client.AMAZE_GUILD_ID
+
     def find_role(self, guild: discord.Guild, search: str) -> discord.Role:
         try:
-            role: discord.Role | None = guild.get_role(int(search))
+            role = guild.get_role(int(search))
             if not role:
                 raise ValueError
         except ValueError:
-            role: discord.Role = match.extractOne(search, guild.roles)[0]
+            role = match.extractOne(search, guild.roles)[0] # type: ignore
 
         return role
 
@@ -63,64 +73,73 @@ class Mod(commands.Cog):
         if not after.guild.id == self.client.AMAZE_GUILD_ID:
             return
 
-        muted_role = after.guild.get_role(self.client.MUTED_ROLE_ID)
-
-        if muted_role not in before.roles and muted_role in after.roles:
-            q = """INSERT INTO stickyroles VALUES ($1, $2)
-                    ON CONFLICT ON CONSTRAINT stickyroles_pkey
-                    DO NOTHING
-                """
-        elif muted_role in before.roles and not muted_role in after.roles:
-            q = """DELETE FROM stickyroles
-                    WHERE user_id = $1 AND role_id = $2
-                """
-        else:
+        if before.roles == after.roles:
             return
 
-        await self.client.db.execute(q, after.id, muted_role.id)
+        added = tuple(r for r in after.roles
+                      if r not in before.roles
+                      and r.id != self.client.MEMBER_ROLE_ID)
+
+        removed = tuple(r for r in before.roles
+                        if r not in after.roles
+                        and r.id != self.client.MEMBER_ROLE_ID)
+
+        if added:
+            args: Tuple[int, ...] = tuple(chain.from_iterable((after.id, r.id) for r in added))
+            await self.client.db.execute(insert_q(len(added)), *args)
+        if removed:
+            q = "DELETE FROM stickyroles WHERE"
+            for i in range(1, len(removed) + 1, 2):
+                base = "" if i == 1 else "OR"
+                q += f"{base} user_id = ${i} AND role_id = ${i + 1} "
+            args: Tuple[int, ...] = tuple(chain.from_iterable((after.id, r.id) for r in removed))
+            await self.client.db.execute(q, *args)
 
     @commands.Cog.listener()
     async def on_member_join(self, member: discord.Member):
-        q = """SELECT role_id FROM stickyroles
-                WHERE user_id = $1
-            """
-        roles: List[discord.Role] = map(
-            lambda rec: member.guild.get_role(rec["role_id"]), await self.client.db.fetch(q, member.id)
-        )
+        if not member.guild.id == self.client.AMAZE_GUILD_ID:
+            return
 
-        await member.add_roles(*roles, reason="sticky roles")
+        extra_roles: Set[discord.Role | None] = {
+            member.guild.get_role(self.client.MEMBER_ROLE_ID)
+        }
+
+        q = """SELECT role_id FROM stickyroles
+               WHERE user_id = $1"""
+        roles: chain[discord.Role] = chain(map( # type: ignore
+            lambda rec: member.guild.get_role(rec["role_id"]), await self.client.db.fetch(q, member.id)
+        ), extra_roles)
+
+        await member.add_roles(*roles, reason="sticky roles", atomic=False)
 
     @commands.Cog.listener()
     async def on_guild_role_delete(self, role: discord.Role):
+        if not role.guild.id == self.client.AMAZE_GUILD_ID:
+            return
+
         q = """DELETE FROM stickyroles
-                WHERE role_id = $1
-            """
+               WHERE role_id = $1"""
         await self.client.db.execute(q, role.id)
 
     @guild_only()
-    async def mute_user(self, interaction: Interaction, target: discord.Member | discord.User):
+    async def mute_user(self, interaction: Interaction, target: discord.Member):
+        assert interaction.guild and isinstance(interaction.user, discord.Member)
         if interaction.user.id == target.id:
             await target.add_roles(
                 discord.Object(self.client.MUTED_ROLE_ID),
                 reason=f"self-mute requested by {interaction.user.name}#{interaction.user.discriminator}"
             )
-            await interaction.response.send_message(
-                self_muted()
-            )
+            await interaction.response.send_message(self_muted())
         elif interaction.guild.get_role(self.client.ADMIN_ROLE_ID) in interaction.user.roles:
             if isinstance(target, discord.Member):
                 await target.add_roles(
                     discord.Object(self.client.MUTED_ROLE_ID),
                     reason=f"member mute requested by {interaction.user.name}#{interaction.user.discriminator}"
                 )
-                return await interaction.response.send_message(
-                    muted(interaction.user)
-                )
-            q = """INSERT INTO stickyroles VALUES ($1, $2)
-                    ON CONFLICT ON CONSTRAINT stickyroles_pkey
-                    DO NOTHING
-                """
-            await self.client.db.execute(q, target.id, self.client.MUTED_ROLE_ID)
+                await interaction.response.defer(ephemeral=True)
+                return await interaction.delete_original_response()
+                #return await interaction.response.send_message(muted(interaction.user))
+            await self.client.db.execute(insert_q(2), target.id, self.client.MUTED_ROLE_ID)
             await interaction.response.send_message(
                 "this user is not in the server, but they've been binded with the sticky role"
             )
@@ -132,26 +151,26 @@ class Mod(commands.Cog):
 
     @commands.command(name="bind", aliases=["bindrole", "br"])
     @commands.is_owner()
-    async def bind_role(self, ctx: commands.Context, target: discord.User | discord.Member, *, role_search: str):
+    @commands.guild_only()
+    async def bind_role(self, ctx: NGKContext, target: discord.User | discord.Member, *, role_search: str):
+        assert ctx.guild is not None
         role = self.find_role(ctx.guild, role_search)
-        q = """INSERT INTO stickyroles VALUES ($1, $2)
-                ON CONFLICT ON CONSTRAINT stickyroles_pkey
-                DO NOTHING
-            """
-        await self.client.db.execute(q, target.id, role.id)
+        if isinstance(target, discord.User):
+            await self.client.db.execute(insert_q(2), target.id, role.id)
 
-        if isinstance(target, discord.Member) and role not in target.roles:
+        elif isinstance(target, discord.Member) and role not in target.roles:
             await target.add_roles(role)
 
         await ctx.message.add_reaction(BotEmojis.YES)
 
     @commands.command(name="unbind", aliases=["unbindrole", "ub"])
     @commands.is_owner()
-    async def unbind_role(self, ctx: commands.Context, target: discord.User, *, role_search: str | None = None):
+    @commands.guild_only()
+    async def unbind_role(self, ctx: NGKContext, target: discord.User, *, role_search: str | None = None):
+        assert ctx.guild is not None
         if not role_search:
             q = """DELETE FROM stickyroles
-                    WHERE user_id = $1
-                """
+                   WHERE user_id = $1"""
             await self.client.db.execute(q, target.id)
             return await ctx.reply(
                 f"all sticky roles cleared for `{target.name}`", mention_author=True
@@ -159,8 +178,7 @@ class Mod(commands.Cog):
 
         role = self.find_role(ctx.guild, role_search)
         q = """DELETE FROM stickyroles
-                WHERE user_id = $1 AND role_id = $2
-            """
+               WHERE user_id = $1 AND role_id = $2"""
         await self.client.db.execute(q, target.id, role.id)
 
         await ctx.message.add_reaction(BotEmojis.YES)
